+++ tests/phpunit/integration/Modules/Analytics_4/SettingsTest.php
@@ -13,6 +13,7 @@ namespace Google\Site_Kit\Tests\Modules\Analytics_4;
 use Google\Site_Kit\Context;
 use Google\Site_Kit\Core\Permissions\Permissions;
 use Google\Site_Kit\Core\Storage\Options;
+use Google\Site_Kit\Modules\Analytics\Settings as Analytics_Settings;
 use Google\Site_Kit\Modules\Analytics_4\Settings;
 use Google\Site_Kit\Tests\Modules\SettingsTestCase;
 
@@ -64,8 +65,6 @@ class SettingsTest extends SettingsTestCase {
 
 		$this->assertEqualSetsWithIndex(
 			array(
-				// TODO: These can be uncommented when Analytics and Analytics 4 modules are officially separated.
-				// 'adsConversionID'        => '',
 				'accountID'                 => '',
 				'adsConversionID'           => '',
 				'adsenseLinked'             => false,
@@ -149,6 +148,45 @@ class SettingsTest extends SettingsTestCase {
 		$this->settings->merge( array( $property_name => $property_value ) );
 
 		$this->assertEquals( $this->user_id, $this->settings->get()['ownerID'] );
+	}
+
+	public function test_retrieve_missing_analytics_4_settings() {
+		delete_option( Settings::OPTION );
+
+		$analytics_settings = new Analytics_Settings( $this->options );
+		$analytics_settings->register();
+		$this->settings->register();
+
+		$options = new Options( new Context( GOOGLESITEKIT_PLUGIN_MAIN_FILE ) );
+		$options->set(
+			Settings::OPTION,
+			array(
+				'propertyID'                => '123',
+				'webDataStreamID'           => '456',
+				'measurementID'             => 'G-789',
+				'useSnippet'                => true,
+				'ownerID'                   => $this->user_id,
+				'googleTagID'               => 'GT-123',
+				'googleTagAccountID'        => '123',
+				'googleTagContainerID'      => '456',
+				'googleTagLastSyncedAtMs'   => 0,
+				'availableCustomDimensions' => null,
+				'propertyCreateTime'        => 0,
+			)
+		);
+
+		$keys_to_check = array(
+			'accountID',
+			'adsConversionID',
+			'adsenseLinked',
+			'canUseSnippet',
+			'trackingDisabled',
+		);
+		$settings      = $options->get( Settings::OPTION );
+
+		foreach ( $keys_to_check as $key ) {
+			$this->assertArrayHasKey( $key, $settings );
+		}
 	}
 
 	public function data_owned_keys() {
